requirements.txt
absl-py==0.11.0
aiohttp==3.7.4
alabaster==0.7.12
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrdict==2.0.1
attrs==20.3.0
audioread==2.1.9
Babel==2.9.0
backcall==0.2.0
black==19.10b0
bleach==3.3.0
boto3==1.17.17
botocore==1.20.17
braceexpand==0.1.6
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.16
editdistance==0.5.3
entrypoints==0.3
eyeD3==0.9.5
filelock==3.0.12
filetype==1.0.7
Flask==1.1.2
frozendict==1.2
fsspec==0.8.7
future==0.18.2
g2p-en==2.1.0
gdown==3.12.2
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.36.0
h5py==3.1.0
hmmlearn==0.2.2
hydra-core==1.0.6
idna==2.10
idna-ssl==1.1.0
imagesize==1.2.0
importlib-metadata==3.7.0
importlib-resources==5.1.1
inflect==5.2.0
iniconfig==1.1.1
ipadic==1.0.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
jieba==0.42.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaldi-io==0.9.4
kaldi-python-io==1.2.1
kaldiio==2.17.2
kiwisolver==1.3.1
latexcodec==2.0.1
librosa==0.8.0
llvmlite==0.35.0
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.10.0
matplotlib==3.1.2
mecab-python3==1.0.3
megatron-lm==1.1.5
mistune==0.8.4
mpmath==1.2.1
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nemo-toolkit==1.0.0rc1
nest-asyncio==1.5.1
nltk==3.5
notebook==6.2.0
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.0.6
onnx==1.8.1
onnxruntime==1.7.0
OpenCC==1.1.1.post1
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
pangu==4.0.6.1
parameterized==0.8.1
parso==0.8.1
pathspec==0.8.1
pathtools==0.1.2
pesq==0.0.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pip-api==0.0.19
pipreqs==0.4.10
plotly==4.1.1
pluggy==0.13.1
pooch==1.3.0
portalocker==2.2.1
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.16
protobuf==3.15.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyannote.core==4.1
pyannote.database==4.0.4
pyannote.metrics==3.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pybtex==0.24.0
pybtex-docutils==1.0.0
pycparser==2.20
pydub==0.25.1
Pygments==2.8.0
pyparsing==2.4.7
pypinyin==0.40.0
pyrsistent==0.17.3
PySocks==1.7.1
pystoi==0.3.3
pytest==6.2.2
pytest-runner==5.3.0
python-dateutil==2.8.1
pytorch-lightning==1.1.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
rapidfuzz==1.1.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
sacrebleu==1.5.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.91
sentry-sdk==0.20.3
shellingham==1.4.0
shortuuid==1.0.1
simplejson==3.16.0
six==1.15.0
smmap==3.0.5
snowballstemmer==2.1.0
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
sox==1.4.1
spectralcluster==0.1.0
SpeechRecognition==3.8.1
Sphinx==3.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.1.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
subprocess32==3.5.4
sympy==1.7.1
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
torch==1.7.1
torch-stft==0.1.4
torchaudio==0.7.2
torchtext==0.8.0
torchvision==0.8.2
tornado==6.1
tqdm==4.46.0
traitlets==4.3.3
transformers==4.5.1
typed-ast==1.4.2
typer==0.3.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.3
wandb==0.10.20
wcwidth==0.2.5
webdataset==0.1.49
webencodings==0.5.1
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yarg==0.1.9
yarl==1.6.3
youtokentome==1.0.6
zipp==3.4.0
